pyproject.toml
[tool.poetry]
name = "bittalk"
version = "0.1.0"
description = ""
authors = ["adver-line <81295787+adver-line@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
django = "^4.2"
black = "^23.3.0"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
